framsticks/framspy/gui/requirements.txt
numpy==1.21.0
Pillow==8.3.1
PyGLM==1.99.3
PyOpenGL==3.1.5
PyOpenGL-accelerate==3.1.5
pyopengltk==0.0.3
